Create AppTitle component which only shows a joke notification on the home page. Otherwise routes to the home page.
